drop go 1.7 support; add go 1.13 support
