+++ gorp_test.go
@@ -2,12 +2,7 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-// Package gorp provides a simple way to marshal Go structs to and from
-// SQL databases.  It uses the database/sql package, and should work with any
-// compliant database/sql driver.
-//
-// Source code and project home:
-// https://github.com/go-gorp/gorp
+// +build integration
 
 package gorp_test
 
@@ -49,9 +44,10 @@ var (
 	debug bool
 )
 
-func init() {
+func TestMain(m *testing.M) {
 	flag.BoolVar(&debug, "trace", true, "Turn on or off database tracing (DbMap.TraceOn)")
 	flag.Parse()
+	os.Exit(m.Run())
 }
 
 type testable interface {
