+++ test_all.sh
@@ -1,4 +1,4 @@
-#!/bin/bash -e
+#!/bin/bash -ex
 
 # on macs, you may need to:
 # export GOBUILDFLAG=-ldflags -linkmode=external
@@ -6,27 +6,27 @@
 coveralls_testflags="-v -covermode=count -coverprofile=coverage.out"
 
 echo "Running unit tests"
-ginkgo -r -race -randomizeAllSpecs -keepGoing -- -test.run TestGorp
+go test -race
 
 echo "Testing against mysql"
 export GORP_TEST_DSN=gorptest/gorptest/gorptest
 export GORP_TEST_DIALECT=mysql
-go test $coveralls_testflags $GOBUILDFLAG $@ .
+go test -tags integration $coveralls_testflags $GOBUILDFLAG $@ .
 
 echo "Testing against gomysql"
 export GORP_TEST_DSN=gorptest:gorptest@/gorptest
 export GORP_TEST_DIALECT=gomysql
-go test $coveralls_testflags $GOBUILDFLAG $@ .
+go test -tags integration $coveralls_testflags $GOBUILDFLAG $@ .
 
 echo "Testing against postgres"
 export GORP_TEST_DSN="user=gorptest password=gorptest dbname=gorptest sslmode=disable"
 export GORP_TEST_DIALECT=postgres
-go test $coveralls_testflags $GOBUILDFLAG $@ .
+go test -tags integration $coveralls_testflags $GOBUILDFLAG $@ .
 
 echo "Testing against sqlite"
 export GORP_TEST_DSN=/tmp/gorptest.bin
 export GORP_TEST_DIALECT=sqlite
-go test $coveralls_testflags $GOBUILDFLAG $@ .
+go test -tags integration $coveralls_testflags $GOBUILDFLAG $@ .
 rm -f /tmp/gorptest.bin
 
 case $(go version) in
