+++ transaction_test.go
@@ -1,3 +1,9 @@
+// Copyright 2012 James Cooper. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+// +build integration
+
 package gorp_test
 
 import "testing"
